tests/unit: refactor stall_detector test functions
Moves some functions and classes defined in the stall_detector test to a
separate header file so they can be reused in other tests.
