+++ tests/unit/stall_detector_test.cc
@@ -35,59 +35,7 @@
 #include <sys/mman.h>
 
 #ifndef SEASTAR_DEBUG
-
-using namespace seastar;
-using namespace std::chrono_literals;
-
-static seastar::logger testlog("testlog");
-
-class temporary_stall_detector_settings {
-    std::chrono::milliseconds _old_threshold;
-    std::function<void ()> _old_report;
-public:
-    /**
-     * Temporarily (until destructor) overload the stall detector threshold and reporting function.
-     *
-     * Also resets the reported stalls counter to zero, so the next backtraces will not be supressed.
-     */
-    temporary_stall_detector_settings(std::chrono::duration<double> threshold, std::function<void ()> report = {})
-            : _old_threshold(engine().get_blocked_reactor_notify_ms())
-            , _old_report(reactor::test::get_stall_detector_report_function()) {
-        engine().update_blocked_reactor_notify_ms(std::chrono::duration_cast<std::chrono::milliseconds>(threshold));
-        reactor::test::set_stall_detector_report_function(std::move(report));
-    }
-
-    ~temporary_stall_detector_settings() {
-        engine().update_blocked_reactor_notify_ms(_old_threshold);
-        reactor::test::set_stall_detector_report_function(std::move(_old_report));
-    }
-};
-
-using void_fn = std::function<void()>;
-
-void spin(std::chrono::duration<double> how_much, void_fn body = []{}) {
-    auto end = internal::cpu_stall_detector::clock_type::now() + how_much;
-    while (internal::cpu_stall_detector::clock_type::now() < end) {
-        body(); // spin!
-    }
-}
-
-static void spin_user_hires(std::chrono::duration<double> how_much) {
-    auto end = std::chrono::high_resolution_clock::now() + how_much;
-    while (std::chrono::high_resolution_clock::now() < end) {
-
-    }
-}
-
-void spin_some_cooperatively(std::chrono::duration<double> how_much, void_fn body = []{}) {
-    auto end = std::chrono::steady_clock::now() + how_much;
-    while (std::chrono::steady_clock::now() < end) {
-        spin(200us, body);
-        if (need_preempt()) {
-            thread::yield();
-        }
-    }
-}
+#include "stall_detector_test_utilities.hh"
 
 SEASTAR_THREAD_TEST_CASE(normal_case) {
     std::atomic<unsigned> reports{};
@@ -131,44 +79,9 @@ SEASTAR_THREAD_TEST_CASE(no_poll_no_stall) {
     BOOST_REQUIRE_EQUAL(reports, 0);
 }
 
-// Triggers stalls by spinning with a specify "body" function
-// which takes most of the spin time.
-static void test_spin_with_body(const char* what, void_fn body) {
-    // The !count_stacks mode outputs stall notification to stderr as usual
-    // and do not assert anything, but are intended for diagnosing
-    // stall problems by inspecting the output. We expect the userspace
-    // spin test to show no kernel callstack, and the kernel test to
-    // show kernel backtraces in the mmap or munmap path, but this is
-    // not exact since neither test spends 100% of its time in the
-    // selected mode (of course, kernel stacks only appear if the
-    // perf-based stall detected could be enabled).
-    //
-    // Then the count_stacks mode tests that the right number of stacks
-    // were output.
-    for (auto count_stacks : {false, true}) {
-        testlog.info("Starting spin test: {}", what);
-        std::atomic<unsigned> reports{};
-        std::function<void()> reporter = count_stacks ? std::function<void()>{[&]{ ++reports; }} : nullptr;
-        temporary_stall_detector_settings tsds(10ms, std::move(reporter));
-        constexpr unsigned nr = 5;
-        for (unsigned i = 0; i < nr; ++i) {
-            spin_some_cooperatively(100ms, body);
-            spin(20ms, body);
-        }
-        testlog.info("Ending spin test: {}", what);
-        BOOST_CHECK_EQUAL(reports, count_stacks ? 5 : 0);
-    }
-}
-
 SEASTAR_THREAD_TEST_CASE(spin_in_userspace) {
     // a body which spends almost all of its time in userspace
     test_spin_with_body("userspace", [] { spin_user_hires(1ms); });
-}
-
-static void mmap_populate(size_t len) {
-    void *p = mmap(nullptr, len, PROT_READ | PROT_WRITE, MAP_PRIVATE | MAP_ANONYMOUS | MAP_POPULATE, 0, 0);
-    BOOST_REQUIRE(p != MAP_FAILED);
-    BOOST_REQUIRE(munmap(p, len) == 0);
 }
 
 SEASTAR_THREAD_TEST_CASE(spin_in_kernel) {
